geospatial-realtime-fleet-telematics/seed/requirements.txt
psycopg2-binary>=2.9.9
shapely>=2.0.0
